requirements.txt
web3>=6.0.0
eth-account>=0.10.0
requests>=2.28.0
python-dotenv>=1.0.0
hyperliquid-python-sdk>=0.4.0
